Completed assignment 1 part 1: Re-architecture package structure; factor out ViewHolder class; enabled orientation changes; used material design guidelines for Dialog; cleaned gradle and updated information; deleted unused methods and imports
