+++ .idea/misc.xml
@@ -1,5 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="DesignSurface">
+    <option name="filePathToZoomLevelMap">
+      <map>
+        <entry key="app/src/main/res/layout/activity_main.xml" value="0.39791666666666664" />
+        <entry key="app/src/main/res/layout/adapter_view.xml" value="0.39791666666666664" />
+        <entry key="app/src/main/res/layout/nav_header.xml" value="0.39791666666666664" />
+        <entry key="app/src/main/res/layout/search_dialog.xml" value="0.3078125" />
+      </map>
+    </option>
+  </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_1_8" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
