+++ app/src/main/AndroidManifest.xml
@@ -14,8 +14,8 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.NasaImageViewer">
 
-        <activity android:name=".MainActivity">
-
+        <activity android:name=".model.MainActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
