+++ app/src/main/java/com/bronzeswordstudios/nasaimageviewer/model/SearchFragment.kt
@@ -1,12 +1,12 @@
-package com.bronzeswordstudios.nasaimageviewer
+package com.bronzeswordstudios.nasaimageviewer.model
 
 import android.app.AlertDialog
 import android.app.Dialog
 import android.content.Context
-import android.content.DialogInterface
 import android.os.Bundle
 import android.widget.EditText
 import androidx.fragment.app.DialogFragment
+import com.bronzeswordstudios.nasaimageviewer.R
 
 class SearchFragment : DialogFragment() {
     private lateinit var listener: SearchDialogListener
@@ -23,8 +23,8 @@ class SearchFragment : DialogFragment() {
         } catch (e: ClassCastException) {
             // The activity must implement the interface or throw an error
             throw ClassCastException(
-                    (context.toString() +
-                            " must implement NoticeDialogListener")
+                (context.toString() +
+                        " must implement NoticeDialogListener")
             )
         }
     }
@@ -40,17 +40,20 @@ class SearchFragment : DialogFragment() {
             // Inflate and set the layout for the dialog - from documentation
             // Pass null as the parent view because its going in the dialog layout - from documentation
             builder.setView(searchView)
-                    // Add action buttons
-                    .setPositiveButton(R.string.submit,
-                            DialogInterface.OnClickListener { dialog, id ->
-                                searchString = searchText.text.toString()
-                                // link the listener to the interface here
-                                listener.onDialogPositiveClick(this)
-                            })
-                    .setNegativeButton(R.string.cancel,
-                            DialogInterface.OnClickListener { dialog, id ->
-                                getDialog()?.cancel()
-                            })
+                .setTitle(R.string.search)
+                // Add action buttons
+                .setPositiveButton(
+                    R.string.submit
+                ) { dialog, id ->
+                    searchString = searchText.text.toString()
+                    // link the listener to the interface here
+                    listener.onDialogPositiveClick(this)
+                }
+                .setNegativeButton(
+                    R.string.cancel
+                ) { dialog, id ->
+                    getDialog()?.cancel()
+                }
             builder.create()
         } ?: throw IllegalStateException("Activity cannot be null")
     }
