+++ app/src/main/java/com/bronzeswordstudios/nasaimageviewer/network/QueryLoader.kt
@@ -1,11 +1,12 @@
-package com.bronzeswordstudios.nasaimageviewer
+package com.bronzeswordstudios.nasaimageviewer.network
 
 import android.content.Context
 import androidx.loader.content.AsyncTaskLoader
+import com.bronzeswordstudios.nasaimageviewer.model.ImageObj
 
 
 class QueryLoader(context: Context, private val url: String) :
-        AsyncTaskLoader<ArrayList<ImageObj>>(context) {
+    AsyncTaskLoader<ArrayList<ImageObj>>(context) {
 
     // load our value on a background thread
     override fun loadInBackground(): ArrayList<ImageObj> {
