+++ app/src/main/res/layout/activity_main.xml
@@ -36,7 +36,7 @@
             android:layout_height="match_parent"
             android:background="@color/white"
             app:layout_behavior="@string/appbar_scrolling_view_behavior"
-            tools:context=".MainActivity">
+            tools:context=".model.MainActivity">
 
             <TextView
                 android:id="@+id/error_view"
