+++ app/src/main/res/layout/search_dialog.xml
@@ -2,17 +2,9 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:orientation="vertical">
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="64dp"
-        android:background="@color/design_default_color_primary_dark"
-        android:fontFamily="sans-serif-medium"
-        android:gravity="center"
-        android:text="@string/search"
-        android:textColor="@color/white"
-        android:textSize="36sp" />
+    android:orientation="vertical"
+    android:paddingStart="24dp"
+    android:paddingEnd="24dp">
 
     <EditText
         android:id="@+id/search_input"
