+++ app/src/main/res/values/themes.xml
@@ -20,4 +20,5 @@
         <item name="android:textColor">#000000</item>
         <item name="fontFamily">Roboto regular</item>
     </style>
+
 </resources>
